Middleware::Logger test and refactor the response
The Rack spec requires the body to respond to #each.
The liter will even fail the test if the body doesn't.
The old implementation was checking to see if the body responded to #body,
and then setting it to `nil` if it didn't, this is likely because the
#body was a Rack::BodyProxy. But the point of a proxy is that it represents
the real thing. It's still subject to the same interface, and as such,
it implements #each. In fact, they even went in and explicitly added that
method when the `method_missing` forwarding they were doing broke on some
Ruby implementations.
https://github.com/rack/rack/blob/ed9050876dfa2418613db05fb52015dc97cb78fb/lib/rack/body_proxy.rb#L26-32

So, checking body is less correct than assuming #each, and thus working with
an Array, or a Rack::BodyProxy, or a Rack::Response (which is what #body returns).

Thus, always iterate through #each, and append its args into an array that we log.

One thing that could cause an issues is if the body can only be iterated once.
If this manifests, probably return the populated array as the new body instead of the original.